+++ Chroma/Chroma/source/Engine/animation/Animator.cpp
@@ -20,7 +20,7 @@ void Animator::PlayTake(std::string const& takeName, float const& normalizedTime
 
 float Animator::CalculateFrameNumber(std::string const& takeName, float const& normalizedTime)
 {
-	return Chroma::Math::Remap(normalizedTime, 0.0, 1.0, 0.0, m_Takes.at(takeName).m_NumFrames);
+	return Chroma::Math::Remap(normalizedTime, 0.0, 1.0, 0.0, (float)m_Takes.at(takeName).m_NumFrames);
 }
 
 void Animator::ApplyAnimJointHierarchy(int const& jointID, KeyFrames& keyFrames, glm::mat4& parentTransform, float const& frameNum)
